fix: wrap package-extraction errors, reject source.method/OLD-side debug-info+devel-pkg
Codex review, fresh evidence on c966f04: _extract_if_package() was called
outside the SnapshotError-to-ClickException boundary, so a recognized-but-
malformed package (e.g. a .deb with no data.tar.* member) leaked a raw
Python traceback instead of the normal concise operational error --
extraction now runs inside the same try/except that already translates
compare_release_against_bundle_facts()'s own SnapshotError/ValueError.
.abicheck.yml's source.method (s1-s6) field is now rejected too, same
root cause as --depth build/source (no channel for L3-L5 evidence
collection). --debug-info old=.../--devel-pkg old=... are now rejected
explicitly -- only the NEW-side scope of each was previously checked,
so an OLD-side operand (which OLD_FACTS, already a resolved snapshot,
has nothing left to apply it to) was silently discarded.

Split compare_bundle_facts.py's growing unsupported-option guard list
into a new sibling module, compare_bundle_facts_rejections.py, once
these additions pushed the dispatcher itself past the architecture
no-growth 800-line cap -- purely a move, no behavior change.
