+++ tests/test_cli_compare_bundle_facts_rejections.py
@@ -23,7 +23,9 @@ rejected *before* any real facts loading/comparison happens, so none of them
 need a real gcc-built bundle -- a placeholder OLD_INPUT/NEW_INPUT pair is
 enough. See ``test_cli_compare_bundle_facts.py``'s own
 ``TestCompareOldBundleFacts`` class for the end-to-end, gcc-built-bundle
-tests this file deliberately doesn't duplicate.
+tests this file deliberately doesn't duplicate, and
+``test_cli_compare_bundle_facts_rejections_more.py`` for this file's own
+continuation once it hit the same test-file cap in a later round.
 """
 
 from __future__ import annotations
@@ -961,140 +963,3 @@ class TestCompareOldBundleFactsEarlyRejections:
 
         assert code == 64
         assert "scope:" in out
-
-    def test_old_side_header_operand_is_rejected(self, tmp_path: Path) -> None:
-        # Codex review: normalize_sided_options puts an old=-scoped --header
-        # into old_headers_only, but _resolve_new_side_headers_includes only
-        # ever reads the new=-scoped/uniform fields -- OLD_FACTS is already
-        # a resolved, stored snapshot with no header re-extraction
-        # available, so this was silently discarded rather than applied.
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("{}")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-        old_header_dir = tmp_path / "old_headers"
-        old_header_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--header",
-            f"old={old_header_dir}",
-            "--format",
-            "json",
-        )
-
-        assert code == 64
-        assert "--header" in out
-
-    def test_old_side_include_operand_is_rejected(self, tmp_path: Path) -> None:
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("{}")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-        old_include_dir = tmp_path / "old_includes"
-        old_include_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--include",
-            f"old={old_include_dir}",
-            "--format",
-            "json",
-        )
-
-        assert code == 64
-        assert "--include" in out
-
-    def test_report_mode_is_rejected(self, tmp_path: Path) -> None:
-        # Codex review: every nested per-library report is rendered via
-        # reporter.to_json(diff) with no report_mode argument -- always the
-        # "full" default, regardless of what was requested.
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("{}")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--report-mode",
-            "leaf",
-            "--format",
-            "json",
-        )
-
-        assert code == 64
-        assert "--report-mode" in out
-
-    def test_show_filtered_is_rejected(self, tmp_path: Path) -> None:
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("{}")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--show-filtered",
-            "--format",
-            "json",
-        )
-
-        assert code == 64
-        assert "--show-filtered" in out
-
-    def test_explicit_jobs_is_rejected(self, tmp_path: Path) -> None:
-        # Codex review: compare_release_against_bundle_facts() processes
-        # every matched library in a synchronous loop -- an explicit
-        # -j/--jobs N request was silently dropped.
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("{}")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--jobs",
-            "4",
-            "--format",
-            "json",
-        )
-
-        assert code == 64
-        assert "--jobs" in out
-
-    def test_default_jobs_is_not_rejected_by_itself(self, tmp_path: Path) -> None:
-        # The silent default (0, "auto-detect") is indistinguishable here
-        # from the flag never having been given, so it must not trip the
-        # --jobs rejection on its own -- confirmed via a malformed OLD_FACTS
-        # document that fails for an unrelated reason first.
-        facts_path = tmp_path / "old.bundlefacts.json"
-        facts_path.write_text("not json")
-        new_dir = tmp_path / "new"
-        new_dir.mkdir()
-
-        code, out = _invoke(
-            "compare",
-            str(facts_path),
-            str(new_dir),
-            "--old-bundle-facts",
-            "--jobs",
-            "0",
-            "--format",
-            "json",
-        )
-
-        assert code != 64 or "--jobs" not in out
